# openapi.yaml
openapi: 3.1.0
info:
  title: VoiceBlender API
  description: >
    VoiceBlender bridges SIP and WebRTC voice calls with multi-party audio
    mixing, real-time speech-to-text, text-to-speech, AI agent integration,
    recording, and webhook-based event delivery.
  version: 1.0.0
  license:
    name: MIT
servers:
  - url: http://localhost:8080/v1
    description: Local development server
...
paths:
...
  /legs/{id}/record:
    parameters:
      - $ref: '#/components/parameters/LegId'
    post:
      operationId: recordLeg
      summary: Start recording a leg to a WAV file
      description: >
        For SIP legs, recording is stereo (left=incoming, right=outgoing).
        For legs in a room, stereo at 16kHz (left=participant audio, right=mixed-minus-self).
      tags: [Legs]
      responses:
        '200':
          description: Recording started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingResponse'
        '404':
          $ref: '#/components/responses/LegNotFound'
        '409':
          description: Leg has no audio reader or room not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to create recording file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      operationId: stopRecordLeg
      summary: Stop recording a leg
      tags: [Legs]
      responses:
        '200':
          description: Recording stopped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingStopResponse'
        '404':
          description: No recording in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /rooms:
    post:
      operationId: createRoom
      summary: Create a room
      tags: [Rooms]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomCreateRequest'
      responses:
        '201':
          description: Room created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '409':
          description: Room ID already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      operationId: listRooms
      summary: List all rooms with participants
      tags: [Rooms]
      responses:
        '200':
          description: Array of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'
  /rooms/{id}:
    parameters:
      - $ref: '#/components/parameters/RoomId'

    get:
      operationId: getRoom
      summary: Get a room with participants
      tags: [Rooms]
      responses:
        '200':
          description: Room details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '404':
          $ref: '#/components/responses/RoomNotFound'

    delete:
      operationId: deleteRoom
      summary: Delete a room
      tags: [Rooms]
      responses:
        '200':
          description: Room deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: deleted
        '404':
          $ref: '#/components/responses/RoomNotFound'
  /rooms/{id}/legs:
    parameters:
      - $ref: '#/components/parameters/RoomId'
    post:
      operationId: addLegToRoom
      summary: Add a leg to a room
      tags: [Rooms]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddLegRequest'
      responses:
        '200':
          description: Leg added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: added
        '400':
          description: Invalid JSON, leg not found, or leg not connected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /rooms/{id}/legs/{legID}:
    parameters:
      - $ref: '#/components/parameters/RoomId'
      - name: legID
        in: path
        required: true
        schema:
          type: string
        description: Leg ID to remove
    delete:
      operationId: removeLegFromRoom
      summary: Remove a leg from a room
      tags: [Rooms]
      responses:
        '200':
          description: Leg removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: removed
        '400':
          description: Room or leg not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /rooms/{id}/record:
    parameters:
      - $ref: '#/components/parameters/RoomId'
    post:
      operationId: recordRoom
      summary: Start recording the room mix to a WAV file
      description: Records the full room mix at 16kHz, 16-bit, mono.
      tags: [Rooms]
      responses:
        '200':
          description: Recording started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingResponse'
        '404':
          $ref: '#/components/responses/RoomNotFound'
        '409':
          description: Room has no participants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to create recording file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      operationId: stopRecordRoom
      summary: Stop room recording
      tags: [Rooms]
      responses:
        '200':
          description: Recording stopped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingStopResponse'
        '404':
          description: No recording in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
    LegId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Leg ID
    RoomId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Room ID
...
  responses:
    LegNotFound:
      description: Leg not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RoomNotFound:
      description: Room not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Leg:
      type: object
      properties:
        leg_id:
          type: string
          description: Unique leg identifier (UUID)
        type:
          type: string
          enum: [sip_inbound, sip_outbound, webrtc]
        state:
          type: string
          enum: [ringing, connected, hung_up]
        room_id:
          type: string
          description: Room ID if the leg is in a room, empty otherwise
      required: [leg_id, type, state]
    Room:
      type: object
      properties:
        id:
          type: string
          description: Room identifier
        participants:
          type: array
          items:
            $ref: '#/components/schemas/Leg'
      required: [id, participants]
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required: [error]
    StatusResponse:
      type: object
      properties:
        status:
          type: string
      required: [status]

    # ── Request bodies ──────────────────────────────────────────────────
...
    RoomCreateRequest:
      type: object
      properties:
        id:
          type: string
          description: Custom room ID (auto-generated UUID if omitted)
    AddLegRequest:
      type: object
      properties:
        leg_id:
          type: string
          description: ID of the leg to add
      required: [leg_id]

    # ── Response bodies ─────────────────────────────────────────────────
...
    RecordingResponse:
      type: object
      properties:
        status:
          type: string
          example: recording
        file:
          type: string
          description: Absolute path to the recording file
      required: [status, file]
    RecordingStopResponse:
      type: object
      properties:
        status:
          type: string
          example: stopped
        file:
          type: string
          description: Absolute path to the recording file
      required: [status, file]

    # ── Webhook events ──────────────────────────────────────────────────
...